Works in Oracle

+++ src/services/src/Demoulas.ProfitSharing.Services/InternalDto/Member.cs
@@ -5,7 +5,7 @@
 /// </summary>
 internal sealed record Member
 {
-    public long Psn { get; init; }
+    public int Psn { get; init; }
     public string? FullName { get; init; } = string.Empty;
     public string FirstName { get; init; } = string.Empty;
     public string? MiddleInitial { get; init; } = string.Empty;
